requirements.txt
faiss-cpu
sentence-transformers
pypdf2
ollama
